Move SPI server core logic into a separate lib
