+++ drv/stm32h7-spi-server-core/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "drv-stm32h7-spi-server-core"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+cfg-if = { workspace = true }
+cortex-m = { workspace = true }
+idol-runtime = { workspace = true }
+num-traits = { workspace = true }
+stm32h7 = { workspace = true }
+zerocopy = { workspace = true }
+
+drv-spi-api = { path = "../spi-api" }
+drv-stm32h7-spi = { path = "../stm32h7-spi" }
+drv-stm32xx-sys-api = { path = "../stm32xx-sys-api" }
+mutable-statics = { path = "../../lib/mutable-statics" }
+ringbuf = { path = "../../lib/ringbuf" }
+userlib = { path = "../../sys/userlib", features = ["panic-messages"] }
+
+[build-dependencies]
+idol = { workspace = true }
+indexmap = { workspace = true }
+proc-macro2 = { workspace = true }
+quote = { workspace = true }
+serde = { workspace = true }
+syn = { workspace = true }
+
+build-util = { path = "../../build/util" }
+call_rustfmt = { path = "../../build/call_rustfmt" }
+
+[features]
+h743 = ["stm32h7/stm32h743", "drv-stm32h7-spi/h743", "drv-stm32xx-sys-api/h743"]
+h753 = ["stm32h7/stm32h753", "drv-stm32h7-spi/h753", "drv-stm32xx-sys-api/h753"]
