+++ drv/stm32h7-spi-server/Cargo.toml
@@ -4,39 +4,29 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-cfg-if = { workspace = true }
-cortex-m = { workspace = true }
 idol-runtime = { workspace = true }
 num-traits = { workspace = true }
-stm32h7 = { workspace = true }
 zerocopy = { workspace = true }
 
 drv-spi-api = { path = "../spi-api" }
-drv-stm32h7-spi = { path = "../stm32h7-spi" }
+drv-stm32h7-spi-server-core = { path = "../stm32h7-spi-server-core" }
 drv-stm32xx-sys-api = { path = "../stm32xx-sys-api" }
-ringbuf = { path = "../../lib/ringbuf" }
 userlib = { path = "../../sys/userlib", features = ["panic-messages"] }
 
 [build-dependencies]
 idol = { workspace = true }
-indexmap = { workspace = true }
-proc-macro2 = { workspace = true }
-quote = { workspace = true }
-serde = { workspace = true }
-syn = { workspace = true }
-
-build-util = { path = "../../build/util" }
-call_rustfmt = { path = "../../build/call_rustfmt" }
 
 [features]
+# These features are unused in the app itself, but Humility uses them to decide
+# which SPI server task to address for a given peripheral.
 spi1 = []
 spi2 = []
 spi3 = []
 spi4 = []
 spi5 = []
 spi6 = []
-h743 = ["stm32h7/stm32h743", "drv-stm32h7-spi/h743", "drv-stm32xx-sys-api/h743"]
-h753 = ["stm32h7/stm32h753", "drv-stm32h7-spi/h753", "drv-stm32xx-sys-api/h753"]
+h743 = ["drv-stm32h7-spi-server-core/h743", "drv-stm32xx-sys-api/h743"]
+h753 = ["drv-stm32h7-spi-server-core/h753", "drv-stm32xx-sys-api/h753"]
 
 # This section is here to discourage RLS/rust-analyzer from doing test builds,
 # since test builds don't work for cross compilation.
